Started implementing custom ego agent model
TODO: currently only the resterized EgoDataset is fed to the EgoAgent's forward function, but the ego agent should receive the vectorized EgoDataset as well as the vectorized AgentDataset. Probably need to code a custom ClosedLoopSimulator in order to be able to feed both datasets as well as using vectorized vs rasterized.